Done with Client and Visitor Pattern!

+++ src/model/visitor/Visitable.java
@@ -1,5 +1,5 @@
 package model.visitor;
 
 public interface Visitable {
-	public float accept(Visitor visitor);
+	public double accept(Visitor visitor);
 }
