+++ src/model/visitor/Visitor.java
@@ -1,5 +1,5 @@
 package model.visitor;
 
 public interface Visitor {
-	public float visit(OrderCart ordercart);//can visit objects in the OrderCart class which has our items
+	public double visit(OrderCart ordercart);//can visit objects in the OrderCart class which has our items
 }
